server/requirements/base.txt
Django==3.1.3
psycopg2-binary==2.8.6
markdown==3.3.3
graphene-django==2.13.0
django-graphql-jwt==0.3.1 
slack-sdk==3.0.0
slack-scim==1.1.0